requirements.txt
requests==2.32.3
apache-airflow-providers-amazon==9.8.0
snowflake-connector-python>=3.0.0
pyyaml>=6.0.0